requirements.txt
aiogram>=3.15
Pillow>=12.0
python-dotenv>=1.1
openai>=1.0
